% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyHistogram.R
\name{safetyHistogram}
\alias{safetyHistogram}
\title{Create a Safety Histogram widget}
\source{
Safety Histogram: \url{https://github.com/RhoInc/safety-histogram}.
}
\usage{
safetyHistogram(data, id_col = "USUBJID", measure_col = "TEST",
  value_col = "STRESN", unit_col = "STRESU", normal_col_low = "STNRLO",
  normal_col_high = "STNRHI", filters_col = NULL, filters_label = NULL,
  details_col = NULL, details_label = NULL, start_value = NULL,
  missingValues = c("", "NA", "N/A"), width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data.}

\item{id_col}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{measure_col}{Name of measure variable name. Default is \code{"TEST"}.}

\item{value_col}{Value of measure variable name. Default is \code{"STRESN"}.}

\item{unit_col}{Unit of measure variable name. Default is \code{"STRESU"}.}

\item{normal_col_low}{Optional: Variable name for column containing lower limit of normal values. Default is \code{"STNRLO"}.}

\item{normal_col_high}{Optional: Variable name for column containing upper limit of normal values. Default is \code{"STNRHI"}.}

\item{filters_col}{Optional vector of variable names to use for filters (in addition to default filter on \code{measure_col}).}

\item{filters_label}{Associated labels to use for filters. If left as \code{NULL}, variable names will be used as labels.}

\item{details_col}{Optional vector of variable names to include in details listing, in addition to variables specified in 
\code{id_col}, \code{value_col}, \code{normal_col_low}, and \code{normal_col_high}.}

\item{details_label}{Associated labels/headers to use for details listing.  If left as \code{NULL}, variable names will be used as labels.}

\item{start_value}{Value of variable defined in \code{measure_col} to be rendered in the histogram when the widget loads.}

\item{missingValues}{Vector of values defining a missing \code{value_col}. Default is \code{c('','NA','N/A')}.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Histogram using R htmlwidgets.
}
\examples{
\dontrun{
# Run Safety Histogram with defaults
safetyHistogram(data=ADBDS)
 
# Run Safety Histogram with some customizations 
safetyHistogram(data=ADBDS, filters_col = c('ARM','SEX','RACE'))
}

}
\seealso{
aeExplorer, aeTimelines, safetyOutlierExplorer, safetyResultsOverTime, safetyShiftPlot
}
