% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyShiftPlot.R
\name{safetyShiftPlot}
\alias{safetyShiftPlot}
\title{Create a Safety Shift Plot widget}
\source{
Safety Shift Plot: \url{https://github.com/RhoInc/safety-shift-plot}.
}
\usage{
safetyShiftPlot(data, id_col = "USUBJID", time_col = "VISITN",
  measure_col = "TEST", value_col = "STRESN", start_value = NULL,
  filters_col = NULL, filters_label = NULL, x_params_visits = NULL,
  x_params_stat = "mean", y_params_visits = NULL, y_params_stat = "mean",
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing the labs data.}

\item{id_col}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{time_col}{Timing of collection variable names. Default is \code{"VISITN"}.}

\item{measure_col}{Name of measure variable name. Default is \code{"TEST"}.}

\item{value_col}{Value of measure variable name. Default is \code{"STRESN"}.}

\item{start_value}{Optional: specifies a value of \code{measure} to be displayed when the chart is loaded.}

\item{filters_col}{Optional vector of variable names to use for filters  (in addition to default filter on \code{measure}).}

\item{filters_label}{Associated labels to use for filters. If left as \code{NULL}, variable names will be used as labels.}

\item{x_params_visits}{Optional character vector of visit values with which to define the baseline outcome. Defaults to the first ordered value of \code{time}.}

\item{x_params_stat}{Aggregate function with which to summarize baseline values.  Default value is \code{mean}. Options are \code{"mean"}, \code{"min"}, or \code{"max"}.}

\item{y_params_visits}{Optional character vector of visit values with which to define the comparison outcome. Defaults to all values of \code{time} except the first sorted value.}

\item{y_params_stat}{Aggregate function with which to summarize comparison values.  Default value is \code{mean}. Options are \code{"mean"}, \code{"min"}, or \code{"max"}.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Shift Plot using R htmlwidgets.
}
\examples{
\dontrun{
# Run Safety Histogram with defaults
safetyShiftPlot(data=ADBDS)

# Run Safety Histogram with some customizations
safetyShiftPlot(data=ADBDS, x_params_visits=1, x_params_stat='min', y_params_visits=c(5,6,7), y_params_stat='mean')
}

}
\seealso{
aeExplorer, aeTimelines, safetyHistogram, safetyOutlierExplorer, safetyResultsOverTime
}
