% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADAE}
\alias{ADAE}
\title{Adverse events sample data}
\format{A data frame with 402 rows and 21 variables:
\describe{
  \item{USUBJID}{Unique participant ID}
  \item{SITEID}{Site ID - numeric}
  \item{SITE}{Site ID - character}
  \item{AGE}{Age}
  \item{SEX}{Sex}
  \item{RACE}{Race}
  \item{ARM}{Adverse event reported term}
  \item{RFSTDTC}{Participant enroll date}
  \item{SAFFL}{Safety population flag}
  \item{ASTDT}{AE start date}
  \item{ASTDY}{Study day of AE start}
  \item{AENDT}{AE end date}
  \item{AENDY}{Study day of AE end}
  \item{AETERM}{Reported term of adverse event}
  \item{AEDECOD}{Adverse event minor category}
  \item{AEBODSYS}{Adverse event major category}
  \item{AESER}{Serious}
  \item{AEONGO}{}
  \item{AESEV}{Severity}
  \item{AEREL}{Relationship}
  \item{AEOUT}{Outcome}
  \item{AESEQ}{Seq number}
  
}}
\source{
\url{https://github.com/RhoInc/viz-library}
}
\usage{
ADAE
}
\description{
A dataset containing simulated adverse events data.  The structure is 1 record per adverse event per participant.
}
\keyword{datasets}
