% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADAE}
\alias{ADAE}
\title{Adverse events sample data}
\format{A data frame with 402 rows and 21 variables:
\describe{
  \item{USUBJID}{Unique participant ID}
  \item{AEBODSYS}{Adverse event major category}
  \item{AEDCOD}{Adverse event minor category}
  \item{ARM}{Adverse event reported term}
  ...
}}
\source{
\url{https://github.com/RhoInc/viz-library}
}
\usage{
ADAE
}
\description{
A dataset containing simulated adverse events data.  The structure is 1 record per adverse event per participant.
}
\keyword{datasets}
