% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADBDS}
\alias{ADBDS}
\title{Safety measures sample data}
\format{A data frame with 22695 rows and 18 variables:
\describe{
  \item{USUBJID}{Unique participant ID}
  \item{TEST}{Measure}
  \item{STRESN}{Value}
  \item{VISITN}{Visit number}
  \item{VISIT}{Visit}
  \item{DY}{Study day}
  \item{DT}{Date}
  \item{AGE}{Age}
  \item{SEX}{Sex}
  \item{RACE}{Race}
  \item{ARM}{Study arm}
  \item{RFSTDTC}{}
  \item{SAFFL}{Safety population}
  \item{CAT}{Specimen}
  \item{STNRLO}{Lower limit of normal}
  \item{STNRHI}{Upper limit of normal}
  \item{STRESU}{Measure units}
  \item{SITEID}{Site ID}
}}
\source{
\url{https://github.com/RhoInc/viz-library}
}
\usage{
ADBDS
}
\description{
A dataset containing simulated safety data.  The structure is 1 record per measure per visit per participant.
}
\keyword{datasets}
