% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ADBDS}
\alias{ADBDS}
\title{Safety measures sample data}
\format{A data frame with 22695 rows and 18 variables:
\describe{
  \item{USUBJID}{Unique participant ID}
  \item{TEST}{Measure}
  \item{VALUE}{Value}
  \item{VISITN}{Timing of collection}
  ...
}}
\source{
\url{https://github.com/RhoInc/viz-library}
}
\usage{
ADBDS
}
\description{
A dataset containing simulated safety data.  The structure is 1 record per measure per visit per participant.
}
\keyword{datasets}
