% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeExplorer.R
\name{aeExplorer}
\alias{aeExplorer}
\title{Create an AE Explorer widget}
\source{
AE Explorer: \url{https://github.com/RhoInc/aeexplorer}.
}
\usage{
aeExplorer(data, id = "USUBJID", major = "AEBODSYS", minor = "AEDECOD",
  group = "ARM", details = NULL, filters_ptcpt_var = NULL,
  filters_ptcpt_label = NULL, filters_event_var = c("AESER", "AESEV",
  "AEREL", "AEOUT"), filters_event_label = c("Serious?", "Severity",
  "Relationship", "Outcome"), groups = NULL, missingValues = c("", "NA",
  "N/A"), filters = NULL, totalCol = TRUE, diffCol = TRUE,
  prefTerms = FALSE, maxPrevalence = 0, maxGroups = 6, h = 1, w = 1,
  r = 1, margin = c(40, 40), diffMargin = c(5, 5), validation = FALSE,
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the Adverse Events data.}

\item{id}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{major}{Higher-level term variable name.  Default is \code{"AEBODSYS"}.}

\item{minor}{Lower-level term variable name.  Default is \code{"AEDECOD"}.}

\item{group}{Group variable name, each value of which displays in its own column in the AE table unless argument \code{groups} is defined. Default is \code{"ARM"}.}

\item{details}{An optional vector of variables to display in the detail listing.  If left as \code{NULL} or unspecified, all variables in input data will appear in detail listing.}

\item{filters_ptcpt_var, filters_ptcpt_label}{Participant-level filters. See details.}

\item{filters_event_var, filters_event_label}{Event-level filters. See details.}

\item{groups}{An option character vector specifying which values to display as columns for variable specified in \code{groups}.  If left as \code{NULL}, all groups will be displayed.}

\item{totalCol}{Specify whether or not to render a total column. Accepts \code{TRUE} (default) or \code{FALSE}.}

\item{diffCol}{Specify whether or not to render a column of graphical differences. Accepts \code{TRUE} (default) or \code{FALSE}.}

\item{prefTerms}{Specify whether or not to initially display all lower-level rows. Accepts \code{"TRUE"} or \code{"FALSE"} (default).}

\item{maxPrevalence}{Filters out any higher- and lower-level rows without at least one group rate above specified value. Default is \code{0}.}

\item{maxGroups}{Number of maximum allowable unique values for variable specified in \code{group}.}

\item{h}{Adjust height of plotted points by adjusting a ratio of the original pixel value. Default is 1 (or 15 px).}

\item{w}{Adjust width of plotted points by adjusting a ratio of the original pixel value. Default is 1 (or 200 px).}

\item{r}{Adjust r of plotted points by adjusting a ratio of the original pixel value. Default is 1 (or 7 px).}

\item{margin}{Numeric vector specfiying the left and right margins for the dot plot.}

\item{diffMargin}{Numeric vector specifying the left and right margins for the difference diamonds plot.}

\item{validation}{Experimental setting that facilitates creating a comma-delimited data set of the current view of the data.  Default is \code{FALSE}.}

\item{width}{Width in pixels.}

\item{height}{Height in pixels.}

\item{elementId}{The element ID for the widget.}

\item{placeholderFlag}{A character vector specifying the value of missing AEs.  Defaults to \code{ c('','NA','N/A')}.}
}
\description{
This function creates an AE Explorer using R htmlwidgets.
}
\details{
\describe{
  \item{filters}{
  There are 4 arguments through which a user can specify filters on a participant and/or event level. 
  At least 1 filter is required. If no label is specified, the variable name will be used.  
  The default is 4 event-level filters.
  }
}
}
\examples{
\dontrun{
# Run AE Explorer with defaults
aeExplorer(data=ADAE)

# Run AE Explore with some customizations 
aeExplorer(data=ADAE, group="ARM", groups=c('Treatment A','Placebo'))
}
   
}
\seealso{
aeTimelines, safetyHistogram, safetyOutlierExplorer, safetyResultsOverTime, safetyShiftPlot
}
