% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeExplorer.R
\name{aeExplorer}
\alias{aeExplorer}
\title{Create an AE Explorer widget}
\source{
AE Explorer: \url{https://github.com/RhoInc/aeexplorer}.
}
\usage{
aeExplorer(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing the Adverse Events data.  Required columns include 
`USUBJID` (unique subject ID), `AEBODSYS` (higher-level term), `AEDECOD` (lower-level term), 
and `ARM` (treatment group),}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates an AE Explorer using R htmlwidgets.
}
\seealso{
aeTimelines, safetyHistogram, safetyOutlierExplorer, safetyResultsOverTime, safetyShiftPlot
}

