% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeTimelines.R
\name{aeTimelines}
\alias{aeTimelines}
\title{Create an AE Timelines widget}
\source{
AE Timelines: \url{https://github.com/RhoInc/ae-timelines}.
}
\usage{
aeTimelines(data, id = "USUBJID", seq = "AESEQ", major = "AEBODSYS",
  minor = "AEDECOD", start = "ASTDY", end = "AENDY", severity = "AESEV",
  related = "AEREL", details = NULL, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the Adverse Events data.}

\item{id}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{seq}{Adverse event sequence number variable name.  Default is \code{"SESEQ"}}

\item{major}{Higher-level term variable name.  Default is \code{"AEBODSYS"}.}

\item{minor}{Lower-level term variable name.  Default is \code{"AEDECOD"}.}

\item{start}{Study day of AE start. Default is \code{"ASTDY"}.}

\item{end}{Study day of AE end. Default is \code{"AENDY"}.}

\item{severity}{Adverse event severity variable name. Default is \code{"AESEV"}.}

\item{related}{Adverse event relatedness to treatment variable name. Default is \code{"AEREL"}}

\item{details}{Optional vector of variable names to include in details listing.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates an AE Timeline using R htmlwidgets.
}
\examples{
\dontrun{
# Run AE Timeline with defaults
aeTimelines(data=ADAE)

# Run AE Timeline with some customizations 
aeTimelines(data=ADAE, details = c('ARM','SEX','RACE'))
}

}
\seealso{
aeExplorer, safetyHistogram, safetyOutlierExplorer, safetyResultsOverTime, safetyShiftPlot
}
