% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aeTimelines.R
\name{aeTimelines}
\alias{aeTimelines}
\title{Create an AE Timelines widget}
\source{
AE Timelines: \url{https://github.com/RhoInc/ae-timelines}.
}
\usage{
aeTimelines(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing the Adverse Events data.

Required columns include 
`USUBJID` (unique subject ID), `AESEQ` (unique sequence ID for each AE), `AEBODSYS` (higher-level term), 
`AETERM` (lower-level term), `ASTDY` (Study day of adverse event start (integer)),
`AENDY` (Study day of adverse event end (integer)), `AESEV` (adeverse event severity), 
and `AEREL` (adverse event relatedness to treatment).}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates an AE Timeline using R htmlwidgets.
}
\seealso{
aeExplorer, safetyHistogram, safetyOutlierExplorer, safetyResultsOverTime, safetyShiftPlot
}

