% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyHistogram.R
\name{safetyHistogram}
\alias{safetyHistogram}
\title{Create a Safety Histogram widget}
\source{
Safety Histogram: \url{https://github.com/RhoInc/safety-histogram}.
}
\usage{
safetyHistogram(data, id = "USUBJID", measure = "TEST", value = "STRESN",
  unit = "STRESU", normal_low = "STRNLO", normal_high = "STRNHI",
  filters = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data.}

\item{id}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{measure}{Name of measure variable name. Default is \code{"TEST"}.}

\item{value}{Value of measure variable name. Default is \code{"STRESN"}.}

\item{unit}{Unit of measure variable name. Default is \code{"STRESU"}.}

\item{normal_low}{Variable name for column containing lower limit of normal values. Default is \code{"STRNLO"}.}

\item{normal_high}{Variable name for column containing upper limit of normal values. Default is \code{"STRNHI"}.}

\item{filters}{An optional list containing individual lists for each desired filter.
To specify a  filter, two arguments (\code{value_col} and \code{label}) are required.   
(e.g. \code{filters = list(list(value_col = "AESER", label = "Serious?"), list(value_col = "SEX", label = "Participant Sex"))})}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Histogram using R htmlwidgets.
}
\examples{
\dontrun{
# Run Safety Histogram with defaults
safetyHistogram(data=LAB)

# Run Safety Histogram with some customizations 
safetyHistogram(data=LAB, filters = list(list(value_col = "RACE", label = "Participant Race"), 
                                         list(value_col = "SEX", label = "Participant Sex")))
}

}
\seealso{
aeExplorer, aeTimelines, safetyOutlierExplorer, safetyResultsOverTime, safetyShiftPlot
}

