% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyHistogram.R
\name{safetyHistogram}
\alias{safetyHistogram}
\title{Create a Safety Histogram widget}
\source{
Safety Histogram: \url{https://github.com/RhoInc/safety-histogram}.
}
\usage{
safetyHistogram(data, id = "USUBJID", measure = "TEST", value = "STRESN",
  unit = "STRESU", normal_low = "STNRLO", normal_high = "STNRHI",
  filters_var = NULL, filters_label = NULL, details_var = NULL,
  details_label = NULL, start_value = NULL, missingValues = c("", "NA",
  "N/A"), width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data.}

\item{id}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{measure}{Name of measure variable name. Default is \code{"TEST"}.}

\item{value}{Value of measure variable name. Default is \code{"STRESN"}.}

\item{unit}{Unit of measure variable name. Default is \code{"STRESU"}.}

\item{normal_low}{Optional: Variable name for column containing lower limit of normal values. Default is \code{"STNRLO"}.}

\item{normal_high}{Optional: Variable name for column containing upper limit of normal values. Default is \code{"STNRHI"}.}

\item{filters_var}{Optional vector of variable names to use for filters (in addition to default filter on \code{measure}).}

\item{filters_label}{Associated labels to use for filters. If left as \code{NULL}, variable names will be used as labels.}

\item{details_var}{Optional vector of variable names to include in details listing, in addition to variables specified in \code{id}, \code{value}, \code{normal_low}, and \code{normal_high}.}

\item{details_label}{Associated labels/headers to use for details listing.  If left as \code{NULL}, variable names will be used as labels.}

\item{start_value}{Value of variable defined in \code{measure} to be rendered in the histogram when the widget loads.}

\item{missingValues}{Vector of values defining a missing \code{value}. Default is \code{c('','NA','N/A')}.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Histogram using R htmlwidgets.
}
\examples{
\dontrun{
# Run Safety Histogram with defaults
safetyHistogram(data=ADBDS)
}

}
\seealso{
aeExplorer, aeTimelines, safetyOutlierExplorer, safetyResultsOverTime, safetyShiftPlot
}
