% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyOutlierExplorer.R
\name{safetyOutlierExplorer}
\alias{safetyOutlierExplorer}
\title{Create a Safety Outlier Explorer widget}
\source{
Safety Outlier Explorer: \url{https://github.com/RhoInc/safety-outlier-explorer}.
}
\usage{
safetyOutlierExplorer(data, id = "USUBJID", timing = c("VISITN", "VISIT",
  "DY"), measure = "TEST", value = "STRESN", unit = "STRESU",
  normal_low = "STNRLO", normal_high = "STNRHI", width = NULL,
  height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data.}

\item{id}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{timing}{Timing of collection variable name(s).  Up to 3 may be specified. Default is \code{c("VISITN","VISIT","DY")}}

\item{measure}{Name of measure variable name. Default is \code{"TEST"}.}

\item{value}{Value of measure variable name. Default is \code{"STRESN"}.}

\item{unit}{Unit of measure variable name. Default is \code{"STRESU"}.}

\item{normal_low}{Optional: Variable name for column containing lower limit of normal values. Default is \code{"STNRLO"}.}

\item{normal_high}{Optional: Variable name for column containing upper limit of normal values. Default is \code{"STNRHI"}.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Outlier Explorer using R htmlwidgets.
}
\examples{
\dontrun{
# Run Safety Outlier Explorer with defaults
safetyOutlierExplorer(data=ADBS) 
}

}
\seealso{
aeExplorer, aeTimelines, safetyHistogram, safetyResultsOverTime, safetyShiftPlot
}
