% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyOutlierExplorer.R
\name{safetyOutlierExplorer}
\alias{safetyOutlierExplorer}
\title{Create a Safety Outlier Explorer widget}
\source{
Safety Outlier Explorer: \url{https://github.com/RhoInc/safety-outlier-explorer}.
}
\usage{
safetyOutlierExplorer(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing the labs data. 

Required columns include `USUBJID` (unique subject ID), 
`VISITN`/`VISIT`/`DY` (timing of collection), `TEST` (name of measure), `STRESN` (value of measure), 
and `STRESU` (unit of measure). 

Optional columns include `STRNLO` (lower limit of normal values) and `STRNHI` (upper limit of normal values).}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Outlier Explorer using R htmlwidgets.
}
\seealso{
aeExplorer, aeTimelines, safetyHistogram, safetyResultsOverTime, safetyShiftPlot
}

