% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyResultsOverTime.R
\name{safetyResultsOverTime}
\alias{safetyResultsOverTime}
\title{Create a Safety Results Over Time widget}
\source{
Safety Results Over Time: \url{https://github.com/RhoInc/safety-results-over-time}.
}
\usage{
safetyResultsOverTime(data, id = "USUBJID", time_var = "VISITN",
  time_label = "Visit Number", time_order = NULL, time_label_rot = FALSE,
  time_label_padding = 0, measure = "TEST", value = "STRESN",
  unit = "STRESU", normal_low = "STNRLO", normal_high = "STNRHI",
  start_value = NULL, groups_var = "NONE", groups_label = "None",
  filters_var = NULL, filters_label = NULL, missingValues = c("", "NA",
  "N/A"), boxplots = TRUE, violins = FALSE, width = NULL, height = NULL,
  elementId = NULL)
}
\arguments{
\item{data}{A data frame containing the labs data.}

\item{id}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{time_var}{Timing of collection variable name. Default is \code{c("VISITN")}.}

\item{time_label}{Label of variable specified in \code{time_var}. Defaults to \code{c('Visit Number')}. If set to \code{NULL}, variable name will be used for labels.}

\item{time_order}{Option to specify order of x-axis using a character vector.}

\item{time_label_rot}{Rotate x-axis tick labels 45 degrees?  Defaults to \code{FALSE}.}

\item{time_label_padding}{X-axis padding for rotated labels in pixels. Defaults to \code{0}.}

\item{measure}{Name of measure variable name. Default is \code{"TEST"}.}

\item{value}{Value of measure variable name. Default is \code{"STRESN"}.}

\item{unit}{Unit of measure variable name. Default is \code{"STRESU"}.}

\item{normal_low}{Variable name for column containing lower limit of normal values. Default is \code{"STNRLO"}.}

\item{normal_high}{Variable name for column containing upper limit of normal values. Default is \code{"STNRHI"}.}

\item{start_value}{Value of variable defined in \code{measure} to be rendered in the plot when the widget loads.}

\item{groups_var}{Optional: name of column(s) to include as options for chart stratification.  Default is \code{"NONE"}.}

\item{groups_label}{Optional label for stratification variable(s).  Default is \code{"None"}. If set to \code{NULL}, variable name will be used as label.}

\item{filters_var}{Optional vector of variable names to use for filters}

\item{filters_label}{Associated labels to use for filters. If left as \code{NULL}, variable names will be used as labels.}

\item{missingValues}{A vector of values found in \code{value} to be ignored when rendering the chat. Default is \code{c('','NA','N/A')}.}

\item{boxplots}{Logical indicating whether to render boxplots.  Default is \code{TRUE}.}

\item{violins}{Logical indicating whether to render violin plots.  Default is \code{FALSE}.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Results Over Time widget using R htmlwidgets.
}
\examples{
\dontrun{
# Run Safety Histogram with defaults
safetyResultsOverTime(data=ADBDS)

# Run Safety Histogram with some customizations 
safetyResultsOverTime(data=ADBDS, groups_var=c("SEX","RACE","ARM"), boxplots=T, violins=T)
}

}
\seealso{
aeExplorer, aeTimelines, safetyHistogram, safetyOutlierExplorer, safetyShiftPlot
}
