% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyResultsOverTime.R
\name{safetyResultsOverTime}
\alias{safetyResultsOverTime}
\title{Create a Safety Results Over Time widget}
\source{
Safety Results Over Time: \url{https://github.com/RhoInc/safety-results-over-time}.
}
\usage{
safetyResultsOverTime(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing the labs data. 

Required columns include `USUBJID` (unique subject ID), 
`VISITN` (timing of collection), `TEST` (name of measure), `STRESN` (value of measure), 
`STRESU` (unit of measure), and `SEX`/`RACE` (subpopulation(s) of interest).

Optional columns include `STRNLO` (lower limit of normal values) and `STRNHI` (upper limit of normal values).}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Results Over Time widget using R htmlwidgets.
}
\seealso{
aeExplorer, aeTimelines, safetyHistogram, safetyOutlierExplorer, safetyShiftPlot
}

