% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyShiftPlot.R
\name{safetyShiftPlot}
\alias{safetyShiftPlot}
\title{Create a Safety Shift Plot widget}
\source{
Safety Shift Plot: \url{https://github.com/RhoInc/safety-shift-plot}.
}
\usage{
safetyShiftPlot(data, id = "USUBJID", timing = "VISITN", measure = "TEST",
  value = "STRESN", start_value = NULL, filters_var = NULL,
  filters_label = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing the labs data.}

\item{id}{Participant ID variable name. Default is \code{"USUBJID"}.}

\item{timing}{Timing of collection variable names. Default is \code{"VISITN"}.}

\item{measure}{Name of measure variable name. Default is \code{"TEST"}.}

\item{value}{Value of measure variable name. Default is \code{"STRESN"}.}

\item{start_value}{Optional: specifies a value of \code{measure} to be displayed when the chart is loaded.}

\item{filters_var}{Optional vector of variable names to use for filters  (in addition to default filter on \code{measure}).}

\item{filters_label}{Associated labels to use for filters. If left as \code{NULL}, variable names will be used as labels.}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Shift Plot using R htmlwidgets.
}
\examples{
\dontrun{
# Run Safety Histogram with defaults
safetyShiftPlot(data=ADBDS)
}

}
\seealso{
aeExplorer, aeTimelines, safetyHistogram, safetyOutlierExplorer, safetyResultsOverTime
}
