% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/safetyShiftPlot.R
\name{safetyShiftPlot}
\alias{safetyShiftPlot}
\title{Create a Safety Shift Plot widget}
\source{
Safety Shift Plot: \url{https://github.com/RhoInc/safety-shift-plot}.
}
\usage{
safetyShiftPlot(data, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{data}{A data.frame containing the labs data. 

Required columns include `USUBJID` (unique subject ID), 
`VISITN` (timing of collection), `TEST` (name of measure), and `STRESN` (value of measure).}

\item{width}{Width in pixels}

\item{height}{Height in pixels}

\item{elementId}{The element ID for the widget.}
}
\description{
This function creates a Safety Shift Plot using R htmlwidgets.
}
\seealso{
aeExplorer, aeTimelines, safetyHistogram, safetyOutlierExplorer, safetyResultsOverTime
}

